move speedometer to component
