+++ components/speedometer/CMakeLists.txt
@@ -1,0 +1,5 @@
+idf_component_register(
+    SRCS "src/speedometer.c"
+    INCLUDE_DIRS "include"
+    REQUIRES system driver
+)
