+++ components/speedometer/idf_component.yml
@@ -1,0 +1,3 @@
+version: "1.0.0"
+description: "speedometer"
+url: "https://github.com/Parahybaja/firmware"
